Changed ExportConfigurationButton setStatus to setConfigured Simplified ButtonProcessing Interface Removed methods from UI class and made them part of the ButtonProcessing
